- Cash deposit for logged user - Re-build of file structure

+++ src/TuserLogged.hpp
@@ -13,7 +13,7 @@ public:
     void checkCurrentBalance(std::string);
     void transactionType();
 
-    void cashDeposit(std::string, double=0.0);
+    void cashDeposit(std::string, int=0);
     void cashWithdraw();
     void cashTransfer();
